requirements.txt
alpaca-py==0.43.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.85.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.2
cryptography==46.0.4
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
Flask==3.1.2
google-api-core==2.29.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-logging==3.13.0
google-cloud-secret-manager==2.26.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==25.0.1
h11==0.16.0
holidays==0.90
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
msgpack==1.1.2
numpy==2.4.2
oauthlib==3.3.1
opentelemetry-api==1.39.1
packaging==26.0
pandas==3.0.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.6
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sseclient-py==1.9.0
starlette==0.52.1
tenacity==9.1.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
Werkzeug==3.1.5
zipp==3.23.0
